Separated test.js into different suites
